refactor: 다른 파일로 분리
- WriteDateInfo
